+++ config/web.jest.config.js
@@ -43,6 +43,7 @@ module.exports = {
   ],
   testPathIgnorePatterns: [
     "/__tests__/__util__/",
+    "/__tests__/__suites__/",
     "/__tests__/__mocks__/",
     "<rootDir>/src/.*/.*(android|ios|native)\.js",
     "<rootDir>/src/index.js",
@@ -52,6 +53,7 @@ module.exports = {
   ,
   coveragePathIgnorePatterns: [
     "/__tests__/__util__/",
+    "/__tests__/__suites__/",
     "<rootDir>/src/.*/.*(android|ios|native)\.js",
     "<rootDir>/src/index.js",
     "<rootDir>/src/init.js",
